fix(widgets): correctness fixes across controls
- ``textbox``: copy/cut called ``substr(begin, end)``, but ``substr``'s second
  argument is a length rather than an end index, so the wrong number of
  characters reached the clipboard. Typed input was cast directly to ``char``,
  truncating any non-ASCII UTF-32 codepoint, and is now encoded with ``utf8()``.
  ``FloatBox::value()`` parsed with ``std::stod``, which throws on empty or
  malformed text, and is replaced by ``std::strtod`` returning zero when nothing
  parses. The float validation pattern also contained ``[e_e]``, a character
  class matching ``e`` or ``_``, where it should accept an ``e``/``E`` exponent.

- ``combobox``: ``scroll_event()`` dismissed the popup and consumed the event on
  every scroll, including horizontal ones with no vertical motion, which are now
  passed through to the base class. ``set_selected_index()`` gained a bounds
  check before indexing into the child buttons.

- ``colorwheel``: ``set_color()`` assigned ``m_hue`` after using it, so the
  white and black components were derived from the previous hue. The HSV members
  were left uninitialized, and ``adjust_position()`` used ``atan`` (which loses
  the quadrant) instead of ``atan2``.

- ``slider``: ``set_value()`` now clamps to the configured range, and drawing
  guards against a zero-width range that would otherwise divide by zero.

- ``tabwidget``: the drag-to-reorder comparison mixed axes, testing ``p.x()``
  against ``m_pos.y()``, and the right-click handler dereferenced the popup even
  when ``m_popup_callback`` returned null.

- ``textarea``: the selection-visible test compared ``m_selection_end`` against
  itself instead of also checking ``m_selection_start``, and the empty-document
  case returned the sentinel ``(-1, 1)`` instead of ``(-1, -1)``.

- ``button``: the ``PopupButton`` ``dynamic_cast`` result was dereferenced
  without a null check, which is invalid for a plain ``Button``.

- ``imageview``/``imagepanel``: ``pixel_to_pos()`` truncated its result to
  integer coordinates, and ``ImagePanel`` could return an index past the end of
  the image list.

- ``common``: ``mainloop_iteration()`` now iterates ``__nanogui_screens`` by
  index, since a draw callback that adds or removes a ``Screen`` would invalidate
  a range-based iterator. ``utf8()`` now rejects codepoints above the Unicode
  maximum of U+10FFFF, and the ``EMSCRIPTEN`` ``run()`` path stores the requested
  ``RunMode`` instead of a removed variable.

+++ src/imageview.cpp
@@ -80,7 +80,7 @@ Vector2f ImageView::pos_to_pixel(const Vector2f &p) const {
 }
 
 Vector2f ImageView::pixel_to_pos(const Vector2f &p) const {
-    Vector2i pos = (p * scale() + m_offset) / screen()->pixel_ratio();
+    Vector2f pos = (p * scale() + m_offset) / screen()->pixel_ratio();
     if (m_draw_border)
         pos += 1.f;
     return pos;
